+++ RDMA/src/include/util/common-structure.h
@@ -4,7 +4,7 @@
 
 #define MAX_SERVER_COUNT 5
 
-typedef int64_t node_id_t;
+typedef uint32_t node_id_t;
 typedef uint32_t req_id_t;
 typedef uint32_t view_id_t;
